cleaned and fixed linear conflict

+++ Sem4/AI/List2/puzzle15/state.hpp
@@ -4,8 +4,6 @@
 #include <cstdint>
 #include <array>
 
-#include <iostream> //tmp
-
 const uint64_t TARGET_STATE = 0x123456789ABCDEF0;
 
 uint64_t convert(std::array<int, 16> arr) {
@@ -16,17 +14,17 @@ uint64_t convert(std::array<int, 16> arr) {
     return state;
 }
 
-inline uint64_t swap(uint64_t state, int z, int x) { // zero pos, x pos
+inline uint64_t swap(uint64_t state, uint8_t z, uint8_t x) { // zero pos, x pos
     return (state | (((state >> (x * 4)) & 0xFULL) << (z * 4))) & (~(0xFULL << (x * 4)));
 }
 
-inline int get_neighbors(uint64_t state, uint64_t neighbors[4]) {
-    int z = 0;
+inline uint8_t get_neighbors(uint64_t state, uint64_t neighbors[4]) {
+    uint8_t z = 0;
     while(((state >> (z * 4)) & 0xFULL) != 0) {
         ++z;
     }
 
-    int c = 0;
+    uint8_t c = 0;
     if(z > 3) { // non-bottom row
         neighbors[c++] = swap(state, z, z - 4);
     }
